+++ CHANGELOG.md
@@ -2,15 +2,16 @@
 
 ## Pre release
 
-## [2019.07.09](https://github.com/uktrade/directory-forms-api/releases/tag/2019.07.09)
-[Full Changelog](https://github.com/uktrade/directory-forms-api/compare/2019.05.02...2019.07.09)
-
 ### Implemented enhancements
 
 - TT-1591 - Expose ingress url to zendesk
 - TT-1160 - Add auto retry to celery tasks
 - No ticket - replace directory components IP whitelister
 - TT-1604 - Send Letter Gov Notify
+- TT-1670 Allow-Govnotify send letters test mode
+
+## [2019.07.09](https://github.com/uktrade/directory-forms-api/releases/tag/2019.07.09)
+[Full Changelog](https://github.com/uktrade/directory-forms-api/compare/2019.05.02...2019.07.09)
 
 ### Fixed bugs: